Partial refactoring of resource handling
