+++ src/resources/font.rs
@@ -1,0 +1,3 @@
+//! Module for actions on a font. 
+//! The underlying type for a font is 
+use rusttype::{self, Font};
