+++ src/resources/image.rs
@@ -1,0 +1,1 @@
+use image;
